extract button component and push down visibility
