+++ client/web-sveltekit/src/routes/search/FileSearchResultHeader.svelte
@@ -3,11 +3,7 @@
 </script>
 
 <script lang="ts">
-    import { mdiContentCopy } from '@mdi/js'
-    import copy from 'copy-to-clipboard'
-
     import { highlightRanges } from '$lib/dom'
-    import Icon from '$lib/Icon.svelte'
     import {
         displayRepoName,
         splitPath,
@@ -17,8 +13,8 @@
         type PathMatch,
         type SymbolMatch,
     } from '$lib/shared'
-    import Tooltip from '$lib/Tooltip.svelte'
-    import { Button } from '$lib/wildcard'
+
+    import CopyPathButton from './CopyPathButton.svelte'
 
     export let result: ContentMatch | PathMatch | SymbolMatch
 
@@ -31,17 +27,6 @@
         result.type !== 'symbol' && result.pathMatches
             ? result.pathMatches.map((match): [number, number] => [match.start.column, match.end.column])
             : []
-
-    let recentlyCopied = false
-    function handleCopyPath(): void {
-        copy(result.path)
-        recentlyCopied = true
-        setTimeout(() => {
-            recentlyCopied = false
-        }, 1000)
-    }
-
-    $: tooltip = recentlyCopied ? 'Copied!' : 'Copy path to clipboard'
 </script>
 
 <a href={repoAtRevisionURL}>{repoName}</a>
@@ -52,14 +37,8 @@
         <a href={fileURL} use:highlightRanges={{ ranges: matches }}>
             {#if fileBase}{fileBase}/{/if}<strong>{fileName}</strong>
         </a>
-        <div class="copy-button">
-            <Tooltip {tooltip} placement="bottom">
-                <Button on:click={() => handleCopyPath()} variant="icon" size="sm" aria-label="Copy path to clipboard">
-                    <Icon inline svgPath={mdiContentCopy} aria-hidden />
-                </Button>
-            </Tooltip>
-        </div>
     {/key}
+    <CopyPathButton path={result.path} />
 </span>
 
 <style lang="scss">
@@ -67,19 +46,5 @@
         display: flex;
         align-items: center;
         gap: 0.5rem;
-    }
-    .copy-button {
-        visibility: hidden;
-        --color: var(--icon-color);
-        &:hover {
-            --color: var(--body-color);
-        }
-    }
-
-    :global(.copy-button-focus-container:hover),
-    :global(.copy-button-focus-container:focus-within) {
-        .copy-button {
-            visibility: visible;
-        }
     }
 </style>
